update: add validation before deleteById, and findByOwnerId

+++ src/main/java/com/rescuewheels/backend/service/VehicleService.java
@@ -3,6 +3,7 @@ package com.rescuewheels.backend.service;
 import com.rescuewheels.backend.dao.VehicleRepository;
 import com.rescuewheels.backend.entity.User;
 import com.rescuewheels.backend.entity.Vehicle;
+import com.rescuewheels.backend.exception.NotUserVehicleException;
 import com.rescuewheels.backend.exception.VehicleNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
@@ -36,16 +37,22 @@ public class VehicleService implements IVehicleService {
 
     @Override
     @Transactional
-    public Vehicle deleteById(String id) {
+    public void deleteById(String id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = (User) authentication.getPrincipal();
         Optional<Vehicle> result = vehicleRepository.findById(id);
 
         if (result.isEmpty()) {
             throw new VehicleNotFoundException("Vehicle id - " + id + " not found");
         }
 
-        vehicleRepository.deleteById(id);
+        Vehicle vehicle = result.get();
 
-        return result.get();
+        if (!vehicle.getOwnerId().equals(user.getId()) && !user.getRoles().contains("ROLE_ADMIN")) {
+            throw new NotUserVehicleException("Vehicle id - " + id + " doesn't belong to user id - " + user.getId());
+        }
+
+        vehicleRepository.deleteById(id);
     }
 
     @Override
@@ -66,6 +73,14 @@ public class VehicleService implements IVehicleService {
 
     @Override
     public List<Vehicle> findByOwnerId(String ownerId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = (User) authentication.getPrincipal();
+
+        if (!ownerId.equals(user.getId()) && !user.getRoles().contains("ROLE_ADMIN")) {
+            throw new NotUserVehicleException("User id - " + user.getId() +
+                    " doesn't have permission to view user id - " + ownerId + " vehicles");
+        }
+
         return vehicleRepository.findByOwnerId(ownerId);
     }
 }
